Move date functions from instruments module to date module
